Add authentication

+++ api/contollers/meeting.js
@@ -1,59 +1,59 @@
 import models from '../models'
+import Sequelize from 'sequelize';
 
-const { meeting, user, room, sequelize, userMeeting } = models
-const include = [
-  {
-    model: user,
-    as: 'users'
-  },
-  {
-    model: room,
-    as: 'room'
-  },
-];
+const { meeting, user, userMeeting, sequelize } = models
+const attributes = { exclude: ['password'] };
 
 export function findAll(req, res) {
-  return meeting.findAll({ include })
-    .then(data => res.send(data))
-    .catch(err => {
-      console.log(err)
-      res.status(400).send(err)
-    })
-}
-
-export function findOne(req, res) {
-  const { id } = req.params;
-
-  return meeting.findByPk(id, { include })
+  return req.current.getMeetings()
     .then(data => res.send(data))
     .catch(err => res.status(400).send(err))
 }
 
+export async function findOne(req, res) {
+  const { id } = req.params;
+
+  try {
+    const dbMeetings = await req.current.getMeetings({ where: { id } });
+    if (!dbMeetings?.length) {
+      return res.status(404).send({ message: 'Meeting not found.'});
+    }
+
+    const m = await meeting.findByPk(id, { include: {
+        model: user,
+        as: 'users',
+        attributes,
+        through: {
+          attributes: []
+        }
+      }
+    });
+    res.send(m);
+  } catch (e) {
+    res.status(400).send(e);
+  }
+}
+
 export async function create(req, res) {
   const { users, ...attributes } = req.body;
+  const currentUser = req.current;
+
   const transaction = await sequelize.transaction();
 
   try {
-    const createdMeeting = await meeting.create(attributes, { transaction });
-
-    users && await Promise.all(users.map(async (item) => {
-      const usr = await user.findByPk(item.id);
-
-      if (!usr) {
-        return res.status(400);
-      }
-
+    const dbMeeting = await meeting.create(attributes, { transaction });
+    await Promise.all([currentUser, ...users.filter(user => user.id !== currentUser.id)].map((item) => {
       const usrMeeting = {
         userId: item.id,
-        meetingId: createdMeeting.id,
+        meetingId: dbMeeting.id,
+        organizer: item.id === currentUser.id
       }
 
       return userMeeting.create(usrMeeting, { transaction });
     }))
-
     await transaction.commit();
 
-    return res.status(200).send(createdMeeting)
+    return res.status(200).send(dbMeeting)
   } catch (e) {
     await transaction.rollback();
     res.status(400).send(e);
@@ -63,21 +63,24 @@ export async function create(req, res) {
 export async function update(req, res) {
   const { id } = req.params;
   const { users, ...attributes } = req.body;
-  const foundMeeting = await meeting.findByPk(id);
-  const usersByMeeting = await foundMeeting.getUsers();
+  const currentUser = req.current;
+
   const transaction = await sequelize.transaction();
 
   try {
-    await foundMeeting.removeUsers(usersByMeeting, { transaction });
-    await Promise.all(users.map((item) => {
-      const usrMeeting = {
-        userId: item.id,
-        meetingId: foundMeeting.id,
-      }
+    const dbMeetings = await req.current.getMeetings({ where: { id }, through: { where: { organizer: true } } });
+    if (!dbMeetings?.length) {
+      return res.status(404).send({ message: 'You are not organizer of this meeting' });
+    }
+    const dbMeeting = dbMeetings[0];
 
-      return userMeeting.create(usrMeeting, { transaction });
-    }))
-
+    const usersByMeetingWithoutOrganizer = await dbMeeting.getUsers({
+      where: { id: { [Sequelize.Op.not]: currentUser.id} }
+    });
+    await dbMeeting.removeUsers(usersByMeetingWithoutOrganizer, { transaction });
+    await Promise.all(users.filter(item => item.id !== currentUser.id).map((item) => (
+      userMeeting.create({ userId: item.id, meetingId: dbMeeting.id }, { transaction })
+    )))
     await meeting.update(attributes, { where: { id }, transaction });
     await transaction.commit();
     return res.status(200).send(req.body)
@@ -90,12 +93,18 @@ export async function update(req, res) {
 export async function del(req, res) {
   const { id } = req.params;
 
-  const foundMeeting = await meeting.findByPk(id);
-  const users = await foundMeeting.getUsers();
   const transaction = await sequelize.transaction();
 
   try {
-    await foundMeeting.removeUsers(users, { transaction });
+    const dbMeetings = await req.current.getMeetings({ where: { id }, through: { where: { organizer: true } } });
+
+    if (!dbMeetings?.length) {
+      return res.status(404).send({ message: 'You are not organizer of this meeting' });
+    }
+
+    const dbMeeting = dbMeetings[0];
+    const users = await dbMeeting.getUsers();
+    await dbMeeting.removeUsers(users, { transaction });
     await meeting.destroy({ where: { id }, transaction });
     await transaction.commit();
     return res.sendStatus(200);
